fix(notifications): preserve stable route fallbacks
