+++ mobile/lib/widgets/profile/profile_header_widget.dart
@@ -42,6 +42,9 @@ import 'package:openvine/widgets/vine_cached_image.dart';
 import 'package:profile_repository/profile_repository.dart';
 import 'package:skeletonizer/skeletonizer.dart';
 
+part 'profile_header_identity.dart';
+part 'profile_header_media.dart';
+
 /// Profile header widget displaying avatar, stats, name, and bio.
 class ProfileHeaderWidget extends ConsumerStatefulWidget {
   const ProfileHeaderWidget({
@@ -425,985 +428,3 @@ class _ProfileHeaderWidgetState extends ConsumerState<ProfileHeaderWidget> {
 /// The username shimmers when an enclosing [Skeletonizer] is enabled;
 /// the NIP-05 / npub identifier and the bio body are wrapped in
 /// [Skeleton.keep] so they stay interactive and unshimmered (#4163).
-class _ProfileNameAndBio extends StatelessWidget {
-  const _ProfileNameAndBio({
-    required this.profile,
-    required this.userIdHex,
-    required this.nip05,
-    required this.about,
-    required this.isOwnProfile,
-    this.displayNameHint,
-    this.accentColor,
-  });
-
-  final UserProfile? profile;
-  final String userIdHex;
-  final String? nip05;
-  final String? about;
-  final bool isOwnProfile;
-  final String? displayNameHint;
-
-  /// Optional accent color (from profile color) for links/buttons.
-  final Color? accentColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        if (profile != null)
-          UserName.fromUserProfile(profile!, style: VineTheme.titleLargeFont())
-        else
-          UserName.fromPubKey(
-            userIdHex,
-            style: VineTheme.titleLargeFont(),
-            anonymousName: displayNameHint,
-          ),
-        Skeleton.keep(
-          child: _UniqueIdentifier(
-            userIdHex: userIdHex,
-            nip05: nip05,
-            isOwnProfile: isOwnProfile,
-            accentColor: accentColor,
-          ),
-        ),
-        Skeleton.keep(child: _ProfileBadgesBlock(userIdHex: userIdHex)),
-        if (about != null && about!.isNotEmpty) ...[
-          const SizedBox(height: 16),
-          Skeleton.keep(child: _AboutText(about: about!)),
-        ],
-        if (profile?.website?.isNotEmpty == true) ...[
-          const SizedBox(height: 8),
-          Skeleton.keep(child: ProfileWebsiteRow(url: profile!.website!)),
-        ],
-        _VerifiedAccountsBlock(isOwnProfile: isOwnProfile),
-      ],
-    );
-  }
-}
-
-class _ProfileBadgesBlock extends ConsumerWidget {
-  const _ProfileBadgesBlock({required this.userIdHex});
-
-  final String userIdHex;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final badges = ref.watch(profileAcceptedBadgesProvider(userIdHex));
-    return badges.when(
-      data: (items) {
-        if (items.isEmpty) return const SizedBox.shrink();
-        return Padding(
-          padding: const EdgeInsets.only(top: 12),
-          child: Wrap(
-            alignment: WrapAlignment.center,
-            spacing: 8,
-            runSpacing: 8,
-            children: [
-              for (final item in items) _ProfileBadgeChip(badge: item),
-            ],
-          ),
-        );
-      },
-      error: (_, _) => const SizedBox.shrink(),
-      loading: () => const SizedBox.shrink(),
-    );
-  }
-}
-
-class _ProfileBadgeChip extends StatelessWidget {
-  const _ProfileBadgeChip({required this.badge});
-
-  final ProfileBadgeViewData badge;
-
-  @override
-  Widget build(BuildContext context) {
-    final imageUrl = badge.imageUrl;
-    final l10n = context.l10n;
-    const radius = 16.0;
-    return Semantics(
-      button: true,
-      label: l10n.profileBadgeSemanticLabel(badge.displayName),
-      child: Material(
-        color: VineTheme.surfaceBackground,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(radius),
-          side: const BorderSide(color: VineTheme.neutral10),
-        ),
-        child: InkWell(
-          borderRadius: BorderRadius.circular(radius),
-          onTap: () => _showProfileBadgeSheet(context, badge),
-          child: ConstrainedBox(
-            constraints: const BoxConstraints(minHeight: 48, minWidth: 48),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  _ProfileBadgeImage(
-                    imageUrl: imageUrl,
-                    semanticLabel: badge.displayName,
-                  ),
-                  const SizedBox(width: 6),
-                  ConstrainedBox(
-                    constraints: const BoxConstraints(maxWidth: 180),
-                    child: Text(
-                      badge.displayName,
-                      style: VineTheme.labelMediumFont(),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ProfileBadgeImage extends StatelessWidget {
-  const _ProfileBadgeImage({
-    required this.imageUrl,
-    this.size = 20,
-    this.semanticLabel,
-  });
-
-  final String? imageUrl;
-  final double size;
-  final String? semanticLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    final label =
-        semanticLabel ?? context.l10n.profileBadgeFallbackSemanticLabel;
-    final fallback = DecoratedBox(
-      decoration: const BoxDecoration(
-        shape: BoxShape.circle,
-        color: VineTheme.vineGreen,
-      ),
-      child: Center(
-        child: ExcludeSemantics(
-          child: Text(
-            'B',
-            style: VineTheme.labelSmallFont(color: VineTheme.primaryDarkGreen),
-          ),
-        ),
-      ),
-    );
-
-    return Semantics(
-      image: true,
-      label: label,
-      child: ExcludeSemantics(
-        child: SizedBox(
-          width: size,
-          height: size,
-          child: imageUrl == null || imageUrl!.isEmpty
-              ? fallback
-              : ClipOval(
-                  child: VineCachedImage(
-                    imageUrl: imageUrl!,
-                    width: size,
-                    height: size,
-                    errorWidget: (_, _, _) => fallback,
-                  ),
-                ),
-        ),
-      ),
-    );
-  }
-}
-
-void _showProfileBadgeSheet(BuildContext context, ProfileBadgeViewData badge) {
-  VineBottomSheet.show<void>(
-    context: context,
-    showHeaderDivider: false,
-    body: _ProfileBadgeDetailsSheet(badge: badge),
-  );
-}
-
-class _ProfileBadgeDetailsSheet extends StatelessWidget {
-  const _ProfileBadgeDetailsSheet({required this.badge});
-
-  static const _maxVisibleRecipients = 12;
-
-  final ProfileBadgeViewData badge;
-
-  @override
-  Widget build(BuildContext context) {
-    final issuerPubkey = badge.issuerPubkey;
-    final recipients = badge.uniqueRecipientPubkeys;
-    final visibleRecipients = recipients.take(_maxVisibleRecipients).toList();
-    final hiddenRecipientCount = recipients.length - visibleRecipients.length;
-    final description = badge.description?.trim();
-    final l10n = context.l10n;
-
-    return SafeArea(
-      top: false,
-      child: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.fromLTRB(20, 8, 20, 24),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  _ProfileBadgeImage(
-                    imageUrl: badge.imageUrl,
-                    size: 56,
-                    semanticLabel: badge.displayName,
-                  ),
-                  const SizedBox(width: 14),
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          badge.displayName,
-                          style: VineTheme.titleMediumFont(
-                            color: VineTheme.onSurface,
-                          ),
-                        ),
-                        if (description != null && description.isNotEmpty) ...[
-                          const SizedBox(height: 6),
-                          Text(
-                            description,
-                            style: VineTheme.bodyMediumFont(
-                              color: VineTheme.onSurfaceVariant,
-                            ),
-                          ),
-                        ],
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-              if (issuerPubkey != null && issuerPubkey.isNotEmpty) ...[
-                const SizedBox(height: 24),
-                _ProfileBadgeSheetSectionTitle(l10n.profileBadgeAwardedBy),
-                UserProfileTile(
-                  pubkey: issuerPubkey,
-                  showFollowButton: false,
-                  onTap: () =>
-                      _openProfileFromBadgeSheet(context, issuerPubkey),
-                ),
-              ],
-              if (recipients.isNotEmpty) ...[
-                const SizedBox(height: 12),
-                _ProfileBadgeSheetSectionTitle(l10n.profileBadgeRecipients),
-                for (final recipientPubkey in visibleRecipients)
-                  UserProfileTile(
-                    pubkey: recipientPubkey,
-                    showFollowButton: false,
-                    onTap: () =>
-                        _openProfileFromBadgeSheet(context, recipientPubkey),
-                  ),
-                if (hiddenRecipientCount > 0)
-                  Padding(
-                    padding: const EdgeInsets.only(left: 16, top: 4),
-                    child: Text(
-                      l10n.profileBadgeMoreRecipients(hiddenRecipientCount),
-                      style: VineTheme.bodySmallFont(
-                        color: VineTheme.onSurfaceVariant,
-                      ),
-                    ),
-                  ),
-              ],
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _ProfileBadgeSheetSectionTitle extends StatelessWidget {
-  const _ProfileBadgeSheetSectionTitle(this.text);
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(left: 16, bottom: 2),
-      child: Text(
-        text,
-        style: VineTheme.labelMediumFont(color: VineTheme.onSurfaceVariant),
-      ),
-    );
-  }
-}
-
-void _openProfileFromBadgeSheet(BuildContext context, String pubkey) {
-  final path = OtherProfileScreen.pathForNpub(
-    NostrKeyUtils.encodePubKey(pubkey),
-  );
-  final router = GoRouter.of(context);
-  Navigator.of(context).pop();
-  router.push(path);
-}
-
-class _VerifiedAccountsBlock extends StatelessWidget {
-  const _VerifiedAccountsBlock({required this.isOwnProfile});
-
-  final bool isOwnProfile;
-
-  @override
-  Widget build(BuildContext context) {
-    final claims = isOwnProfile
-        ? _readMyClaims(context)
-        : _readOtherClaims(context);
-    if (claims.isEmpty) return const SizedBox.shrink();
-    return Padding(
-      padding: const EdgeInsets.only(top: 12),
-      child: VerifiedAccountsRow(claims: claims),
-    );
-  }
-
-  static List<IdentityClaim> _readMyClaims(BuildContext context) {
-    try {
-      return context.select<MyProfileBloc, List<IdentityClaim>>((bloc) {
-        final state = bloc.state;
-        if (state is MyProfileLoaded) return state.verifiedClaims;
-        if (state is MyProfileUpdated) return state.verifiedClaims;
-        return const [];
-      });
-    } on ProviderNotFoundException {
-      return const [];
-    }
-  }
-
-  static List<IdentityClaim> _readOtherClaims(BuildContext context) {
-    try {
-      return context.select<OtherProfileBloc, List<IdentityClaim>>((bloc) {
-        final state = bloc.state;
-        if (state is OtherProfileLoaded) return state.verifiedClaims;
-        return const [];
-      });
-    } on ProviderNotFoundException {
-      return const [];
-    }
-  }
-}
-
-/// Unique identifier display (NIP-05 or full npub with ellipsis).
-/// Uses profile accent color when available, falls back to vineGreen.
-/// Shows warning for failed NIP-05 verification on own profile.
-/// Hides unverified NIP-05s for other profiles (potential impersonation).
-class _UniqueIdentifier extends ConsumerWidget {
-  const _UniqueIdentifier({
-    required this.userIdHex,
-    required this.nip05,
-    required this.isOwnProfile,
-    this.accentColor,
-  });
-
-  final String userIdHex;
-  final String? nip05;
-  final bool isOwnProfile;
-
-  /// Optional accent color (from profile color) for the link text and icon.
-  final Color? accentColor;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final hasNip05 = nip05 != null && nip05!.isNotEmpty;
-    final npub = NostrKeyUtils.encodePubKey(userIdHex);
-    const linkColor = VineTheme.vineGreen;
-
-    // Watch NIP-05 verification status
-    final verificationStatus = hasNip05
-        ? ref
-              .watch(nip05VerificationProvider(userIdHex))
-              .whenOrNull(data: (status) => status)
-        : null;
-
-    final verificationFailed =
-        verificationStatus == Nip05VerificationStatus.failed;
-
-    // For other profiles: hide unverified NIP-05s (show npub instead)
-    // For own profile: show with warning so user knows there's an issue
-    final String displayText;
-    if (hasNip05) {
-      if (verificationFailed && !isOwnProfile) {
-        // Don't show unverified NIP-05s for other users - potential impersonation
-        displayText = truncateNpubForDisplay(npub);
-      } else {
-        displayText = nip05!;
-      }
-    } else {
-      displayText = truncateNpubForDisplay(npub);
-    }
-
-    return GestureDetector(
-      onTap: () {
-        final verifiedNip05 = hasNip05 && !verificationFailed ? nip05 : null;
-        final profileUrl = buildProfileUrl(verifiedNip05, npub);
-        ClipboardUtils.copy(
-          context,
-          profileUrl,
-          message: context.l10n.profileLinkCopied,
-        );
-      },
-      child: Text(
-        displayText,
-        style: VineTheme.bodyMediumFont(color: linkColor),
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-      ),
-    );
-  }
-}
-
-/// Maximum characters of a raw npub to show before the ellipsis when it is
-/// used as the fallback identifier on the profile screen.
-const int profileNpubMaxChars = 16;
-
-/// Trim a raw npub for the profile identifier row.
-///
-/// Shows the first [profileNpubMaxChars] characters followed by an ellipsis
-/// when the npub is longer than that. Used only when no NIP-05 / divine
-/// username is available (or the NIP-05 is unverified on another user's
-/// profile).
-String truncateNpubForDisplay(String npub) {
-  if (npub.length <= profileNpubMaxChars) return npub;
-  return '${npub.substring(0, profileNpubMaxChars)}...';
-}
-
-/// Build a shareable profile URL.
-///
-/// If the user has a `.divine.video` NIP-05 subdomain (e.g. `_@thomas.divine.video`),
-/// returns `https://thomas.divine.video`. Otherwise falls back to
-/// `https://divine.video/profile/{npub}`.
-@visibleForTesting
-String buildProfileUrl(String? nip05, String npub) {
-  if (nip05 != null && nip05.isNotEmpty) {
-    // NIP-05 format: `_@username.divine.video` or `user@domain.com`
-    final atIndex = nip05.indexOf('@');
-    if (atIndex != -1) {
-      final domain = nip05.substring(atIndex + 1);
-      if (domain.endsWith('.divine.video')) {
-        return 'https://$domain';
-      }
-    }
-  }
-  return 'https://divine.video/profile/$npub';
-}
-
-/// About/bio text display with expandable "Show more/less" functionality.
-class _AboutText extends StatefulWidget {
-  const _AboutText({required this.about});
-
-  final String about;
-
-  /// Maximum lines to show when collapsed.
-  static const int _collapsedMaxLines = 3;
-
-  @override
-  State<_AboutText> createState() => _AboutTextState();
-}
-
-class _AboutTextState extends State<_AboutText> {
-  bool _isExpanded = false;
-  bool _needsExpansion = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final textStyle = VineTheme.bodyMediumFont(
-      color: VineTheme.onSurfaceVariant,
-    );
-
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        // Measure if text exceeds max lines
-        final textSpan = TextSpan(text: widget.about, style: textStyle);
-        final textPainter = TextPainter(
-          text: textSpan,
-          maxLines: _AboutText._collapsedMaxLines,
-          textDirection: TextDirection.ltr,
-        )..layout(maxWidth: constraints.maxWidth);
-
-        _needsExpansion = textPainter.didExceedMaxLines;
-
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (_isExpanded)
-              SelectableLinkifiedText(text: widget.about, style: textStyle)
-            else
-              LinkifiedText(
-                text: widget.about,
-                style: textStyle,
-                maxLines: _AboutText._collapsedMaxLines,
-                overflow: TextOverflow.ellipsis,
-              ),
-            if (_needsExpansion)
-              GestureDetector(
-                onTap: () => setState(() => _isExpanded = !_isExpanded),
-                child: Padding(
-                  padding: const EdgeInsets.only(top: 4),
-                  child: Text(
-                    _isExpanded
-                        ? context.l10n.profileShowLess
-                        : context.l10n.profileShowMore,
-                    style: VineTheme.bodySmallFont(color: VineTheme.vineGreen),
-                  ),
-                ),
-              ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-/// The 334px background banner area. Shows a banner image, a color gradient,
-/// or a plain dark background depending on what the profile provides.
-///
-/// Uses a foreground [BoxDecoration] gradient scrim instead of a [Stack] to
-/// avoid parentData assertions during route transitions.
-class ProfileBanner extends StatelessWidget {
-  const ProfileBanner({
-    required this.height,
-    this.bannerUrl,
-    this.profileColor,
-    super.key,
-  });
-
-  final double height;
-  final String? bannerUrl;
-  final Color? profileColor;
-
-  @override
-  Widget build(BuildContext context) {
-    // Gradient scrim that fades to the surface background at the bottom.
-    // Applied as foregroundDecoration so it overlays the background content.
-    final scrimDecoration = BoxDecoration(
-      gradient: LinearGradient(
-        begin: Alignment.topCenter,
-        end: Alignment.bottomCenter,
-        colors: [
-          VineTheme.surfaceBackground.withValues(alpha: 0),
-          VineTheme.surfaceBackground,
-        ],
-      ),
-    );
-
-    if (bannerUrl != null) {
-      return _BannerImage(
-        bannerUrl: bannerUrl!,
-        height: height,
-        scrimDecoration: scrimDecoration,
-      );
-    }
-
-    final backgroundGradient = profileColor != null
-        ? LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [profileColor!, profileColor!],
-          )
-        : const LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [VineTheme.containerLow, VineTheme.surfaceBackground],
-          );
-
-    return Container(
-      width: double.infinity,
-      height: height,
-      decoration: BoxDecoration(gradient: backgroundGradient),
-      foregroundDecoration: scrimDecoration,
-    );
-  }
-}
-
-/// Banner with a network image and a gradient scrim overlay.
-/// Uses [VineCachedImage] so banner art benefits from the shared on-disk
-/// cache layer instead of hitting the network on every rebuild.
-class _BannerImage extends StatelessWidget {
-  const _BannerImage({
-    required this.bannerUrl,
-    required this.height,
-    required this.scrimDecoration,
-  });
-
-  final String bannerUrl;
-  final double height;
-  final BoxDecoration scrimDecoration;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: double.infinity,
-      height: height,
-      foregroundDecoration: scrimDecoration,
-      clipBehavior: Clip.hardEdge,
-      decoration: const BoxDecoration(color: VineTheme.surfaceBackground),
-      child: VineCachedImage(
-        imageUrl: bannerUrl,
-        errorWidget: (_, _, _) => const SizedBox.shrink(),
-      ),
-    );
-  }
-}
-
-/// Stats row displaying Followers, Following, Likes, and Loops with dividers.
-///
-/// Shows a skeleton for up to [_ProfileStatsRowState._skeletonTimeout] while
-/// stats are being fetched. After the timeout the row keeps all four columns
-/// visible but renders a `—` placeholder for each count, rather than
-/// shimmering indefinitely or collapsing the row (which would shift the
-/// surrounding profile layout).
-class _ProfileStatsRow extends StatefulWidget {
-  const _ProfileStatsRow({required this.userIdHex, this.profileStats});
-
-  final String userIdHex;
-  final ProfileStats? profileStats;
-
-  @override
-  State<_ProfileStatsRow> createState() => _ProfileStatsRowState();
-}
-
-class _ProfileStatsRowState extends State<_ProfileStatsRow> {
-  static const _skeletonTimeout = Duration(seconds: 7);
-
-  /// Two-digit placeholder painted behind the Skeletonizer shimmer while the
-  /// real stats load. The number itself is never visible — it only sets the
-  /// width of the skeleton bar.
-  static const _skeletonPlaceholderCount = 99;
-
-  Timer? _timer;
-  bool _timeoutExpired = false;
-
-  @override
-  void initState() {
-    super.initState();
-    if (widget.profileStats == null) {
-      _timer = Timer(_skeletonTimeout, () {
-        if (mounted) setState(() => _timeoutExpired = true);
-      });
-    }
-  }
-
-  @override
-  void dispose() {
-    _timer?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isLoading = widget.profileStats == null;
-
-    final hasFollowers = widget.profileStats?.followers != null;
-    final hasFollowing = widget.profileStats?.following != null;
-    final hasLikes = widget.profileStats?.totalLikes != null;
-    final hasLoops = widget.profileStats?.totalViews != null;
-
-    final l10n = context.l10n;
-    final columns = <Widget>[
-      if (hasLoops || isLoading)
-        ProfileStatColumn(
-          count: isLoading
-              ? _skeletonPlaceholderCount
-              : widget.profileStats!.totalViews,
-          label: l10n.profileLoopsLabel,
-          isLoading: isLoading && _timeoutExpired,
-        ),
-      if (hasLikes || isLoading)
-        ProfileStatColumn(
-          count: isLoading
-              ? _skeletonPlaceholderCount
-              : widget.profileStats!.totalLikes,
-          label: l10n.profileLikesLabel,
-          isLoading: isLoading && _timeoutExpired,
-        ),
-      if (hasFollowing || isLoading)
-        ProfileStatColumn(
-          count: isLoading
-              ? _skeletonPlaceholderCount
-              : widget.profileStats!.following,
-          label: l10n.profileFollowingLabel,
-          isLoading: isLoading && _timeoutExpired,
-          onTap: () => context.push(
-            FollowingScreenRouter.pathForPubkey(widget.userIdHex),
-          ),
-        ),
-      if (hasFollowers || isLoading)
-        ProfileStatColumn(
-          count: isLoading
-              ? _skeletonPlaceholderCount
-              : widget.profileStats!.followers,
-          label: l10n.profileFollowersLabel,
-          isLoading: isLoading && _timeoutExpired,
-          onTap: () => context.push(
-            FollowersScreenRouter.pathForPubkey(widget.userIdHex),
-          ),
-        ),
-    ];
-
-    return Skeletonizer(
-      enabled: isLoading && !_timeoutExpired,
-      enableSwitchAnimation: true,
-      effect: vineSkeletonEffect,
-      child: Row(
-        children: [
-          for (int i = 0; i < columns.length; i++) ...[
-            if (i > 0) const _StatDivider(),
-            Expanded(child: columns[i]),
-          ],
-        ],
-      ),
-    );
-  }
-}
-
-/// Vertical divider between stat columns.
-class _StatDivider extends StatelessWidget {
-  const _StatDivider();
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox(
-      width: 2,
-      height: 44,
-      child: ColoredBox(color: VineTheme.outlineMuted),
-    );
-  }
-}
-
-/// Profile avatar with optional action label overlapping the bottom edge.
-///
-/// When [pendingActions] is non-empty, a pill-shaped label (e.g. "Secure
-/// your account") is centred below the avatar. A red count badge appears
-/// on the label when more than one action is pending. Tapping either the
-/// avatar or the label triggers [onActionTap].
-///
-/// The avatar shimmers when an enclosing [Skeletonizer] is enabled. The
-/// pending-action label is decorative chrome and is wrapped in
-/// [Skeleton.keep] so it stays interactive and unshimmered (#4163).
-class _ProfileAvatarWithColor extends StatelessWidget {
-  const _ProfileAvatarWithColor({
-    required this.imageUrl,
-    required this.userIdHex,
-    this.profileColor,
-    this.pendingActions = const [],
-    this.onActionTap,
-  });
-
-  final String? imageUrl;
-
-  /// Hex pubkey used as the placeholder tone seed so the same user gets
-  /// the same accent color here as in notifications and other surfaces.
-  final String userIdHex;
-  final Color? profileColor;
-
-  /// Ordered list of pending profile actions. The first action determines
-  /// the label text and icon.
-  final List<ProfileActionType> pendingActions;
-
-  /// Called when the label or avatar is tapped.
-  final VoidCallback? onActionTap;
-
-  @override
-  Widget build(BuildContext context) {
-    const avatarSize = 144.0;
-    final hasAvatar = imageUrl != null && imageUrl!.isNotEmpty;
-    final avatarWidget = UserAvatar(
-      imageUrl: imageUrl,
-      placeholderSeed: userIdHex,
-      size: avatarSize,
-    );
-    final avatar = hasAvatar
-        ? GestureDetector(
-            onTap: () => _showAvatarLightbox(
-              context,
-              imageUrl: imageUrl,
-              userIdHex: userIdHex,
-            ),
-            child: avatarWidget,
-          )
-        : avatarWidget;
-
-    if (pendingActions.isEmpty) return avatar;
-
-    // The label overlaps the avatar bottom, so we need extra space below.
-    return Stack(
-      clipBehavior: Clip.none,
-      alignment: Alignment.topCenter,
-      children: [
-        // Reserve space for label overflow below the avatar.
-        const SizedBox(width: avatarSize, height: avatarSize + 16),
-        avatar,
-        Positioned(
-          bottom: 0,
-          child: Skeleton.replace(
-            replacement: const SizedBox.shrink(),
-            child: GestureDetector(
-              onTap: onActionTap,
-              child: _ProfileActionLabel(
-                action: pendingActions.first,
-                badgeCount: pendingActions.length,
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-/// Pill-shaped label shown below the profile avatar for the highest-priority
-/// pending action. Displays an icon and text matching the action type, with
-/// an optional red badge when multiple actions are pending.
-class _ProfileActionLabel extends StatelessWidget {
-  const _ProfileActionLabel({required this.action, required this.badgeCount});
-
-  final ProfileActionType action;
-  final int badgeCount;
-
-  @override
-  Widget build(BuildContext context) {
-    final (icon, label) = switch (action) {
-      ProfileActionType.secureAccount => (
-        DivineIconName.lockSimple,
-        'Secure your account',
-      ),
-      ProfileActionType.completeProfile => (
-        DivineIconName.pencilSimple,
-        'Complete your profile',
-      ),
-    };
-
-    return Stack(
-      clipBehavior: Clip.none,
-      children: [
-        Container(
-          padding: const EdgeInsets.fromLTRB(12, 8, 16, 8),
-          decoration: BoxDecoration(
-            color: VineTheme.accentYellowBackground,
-            borderRadius: BorderRadius.circular(16),
-            boxShadow: const [
-              BoxShadow(
-                color: Color(0x1A000000),
-                offset: Offset(0.4, 0.4),
-                blurRadius: 0.6,
-              ),
-              BoxShadow(
-                color: Color(0x1A000000),
-                offset: Offset(1, 1),
-                blurRadius: 1,
-              ),
-            ],
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            spacing: 8,
-            children: [
-              DivineIcon(icon: icon, size: 16, color: VineTheme.accentYellow),
-              Text(
-                label,
-                style: VineTheme.titleSmallFont(color: VineTheme.accentYellow),
-              ),
-            ],
-          ),
-        ),
-        if (badgeCount > 0)
-          Positioned(
-            right: -4,
-            top: -8,
-            child: Container(
-              constraints: const BoxConstraints(minWidth: 20, minHeight: 20),
-              padding: const EdgeInsets.symmetric(horizontal: 4),
-              decoration: BoxDecoration(
-                color: VineTheme.error,
-                borderRadius: BorderRadius.circular(1000),
-              ),
-              alignment: Alignment.center,
-              child: Text(
-                badgeCount.toString(),
-                style: VineTheme.labelSmallFont(),
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Avatar lightbox
-// ---------------------------------------------------------------------------
-
-void _showAvatarLightbox(
-  BuildContext context, {
-  required String userIdHex,
-  String? imageUrl,
-}) {
-  showGeneralDialog<void>(
-    context: context,
-    barrierColor: VineTheme.transparent,
-    barrierDismissible: true,
-    barrierLabel: context.l10n.profileAvatarLightboxBarrierLabel,
-    pageBuilder: (context, _, _) =>
-        _AvatarLightbox(imageUrl: imageUrl, userIdHex: userIdHex),
-  );
-}
-
-class _AvatarLightbox extends StatelessWidget {
-  const _AvatarLightbox({required this.userIdHex, this.imageUrl});
-
-  final String? imageUrl;
-
-  /// Pubkey used as the placeholder seed so the lightbox's fallback
-  /// colour matches the avatar everywhere else when the image fails.
-  final String userIdHex;
-
-  @override
-  Widget build(BuildContext context) {
-    final safeAreaTop = MediaQuery.of(context).padding.top;
-
-    return Semantics(
-      label: context.l10n.profileAvatarLightboxCloseSemanticLabel,
-      button: true,
-      child: GestureDetector(
-        onTap: () => Navigator.of(context).pop(),
-        child: SizedBox.expand(
-          child: BackdropFilter(
-            filter: ImageFilter.blur(sigmaX: 40, sigmaY: 40),
-            child: ColoredBox(
-              color: VineTheme.scrim65,
-              child: Stack(
-                children: [
-                  Center(
-                    child: UserAvatar(
-                      imageUrl: imageUrl,
-                      placeholderSeed: userIdHex,
-                      size: 288,
-                      cornerRadius: 112,
-                    ),
-                  ),
-                  Positioned(
-                    top: safeAreaTop + 12,
-                    left: 12,
-                    child: DivineIconButton(
-                      icon: DivineIconName.x,
-                      type: DivineIconButtonType.ghostSecondary,
-                      size: DivineIconButtonSize.small,
-                      onPressed: () => Navigator.of(context).pop(),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
