+++ mobile/scripts/baseline/file_sizes.txt
@@ -11,13 +11,13 @@ lib/main.dart	2715
 lib/providers/clip_manager_provider.dart	894
 lib/providers/individual_video_providers.dart	1410
 lib/providers/video_editor_provider.dart	1051
-lib/router/app_router.dart	1570
+lib/router/app_router.dart	1562
 lib/screens/apps/nostr_app_sandbox_screen.dart	842
 lib/screens/auth/invite_gate_screen.dart	819
 lib/screens/auth/nostr_connect_screen.dart	817
 lib/screens/creator_analytics_screen.dart	1406
-lib/screens/explore_screen.dart	1282
-lib/screens/feed/pooled_fullscreen_video_feed_screen.dart	1063
+lib/screens/explore_screen.dart	1258
+lib/screens/feed/pooled_fullscreen_video_feed_screen.dart	1055
 lib/screens/library_screen.dart	884
 lib/screens/profile_setup_screen.dart	1897
 lib/screens/relay_diagnostic_screen.dart	1228
@@ -30,11 +30,10 @@ lib/services/curated_list_service.dart	1785
 lib/services/notification_service.dart	861
 lib/services/upload_manager.dart	2854
 lib/services/video_editor/video_editor_render_service.dart	1065
-lib/services/video_event_publisher.dart	1848
+lib/services/video_event_publisher.dart	1709
 lib/services/video_event_service.dart	6034
 lib/services/video_publish/video_publish_service.dart	843
 lib/services/zendesk_support_service.dart	1168
-lib/widgets/profile/profile_header_widget.dart	1141
 lib/widgets/report_content_dialog.dart	857
 lib/widgets/share_video_menu.dart	1606
 lib/widgets/user_picker_sheet.dart	1009
